requirements.txt
nose
sphinx
django-datatable-view
ply